feat: integrate current user data into profile page and improve project overview
- Updated ProfilePage to fetch current user data using useCurrentUser hook.
- Refactored formData initialization to use currentUser's displayName and avatar.
- Simplified ProjectPage by dynamically importing ProjectOverview and removing unused states and loading indicators.
- Enhanced SprintPage to dynamically import SprintBoardDnd and handle loading states more efficiently.
- Added RecentActivity component to display user activities in a structured format.
- Improved StatusOverview to display project statistics with better styling and data handling.
- Updated various components to use 'use client' directive for proper client-side rendering.
- Refactored project and sprint related components to ensure data is fetched and displayed correctly.
- Changed sprint date types in typescript definitions from string to Date for better date handling. #15
