+++ src/app/(dashboard)/projects/components/CreateProjectModal.tsx
@@ -1,3 +1,4 @@
+'use client'
 import { useState, useEffect } from 'react'
 import { useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
