+++ src/components/projects/edit-task-modal.tsx
@@ -1,3 +1,4 @@
+'use client';
 import { useState } from 'react';
 import { Task, BoardColumn } from '@/lib/types';
 import {
