+++ src/components/projects/invite-team-modal.tsx
@@ -1,3 +1,4 @@
+'use client';
 import { useState } from 'react';
 import {
   Dialog,
